Added method="auto" to download.file calls
R win-builder package build failed with error on download of file from
RBA website and issued warning on CRAN web address (https://...), but
does not fail on local Linux and Windows machines.

Could be a download method error---e.g. according to
https://stat.ethz.ch/pipermail/r-package-devel/2018q1/002485.html, the
Winbuilder machine uses WinHTTP settings that correspond to IE9.
This change to test if that resolves issue.